+++ src/components/atoms/index.ts
@@ -1,4 +1,4 @@
 export { default as AGamepad } from './AGamepad.vue';
 export { default as AMessage } from './AMessage.vue';
 export { default as ATextLogo } from './ATextLogo.vue';
-export { default as AAnnounce } from './AAnnounce.vue';
+export { default as AAnnouncement } from './AAnnouncement.vue';
