+++ src/components/molecules/index.ts
@@ -1,9 +1,9 @@
 export { default as MHeader } from './MHeader.vue';
 export { default as MFooter } from './MFooter.vue';
 export { default as MPattern } from './MPattern.vue';
-export { default as MAnnounce } from './MAnnounce.vue';
 export { default as MToolsMenu } from './MToolsMenu.vue';
 export { default as MHeaderMenu } from './MHeaderMenu.vue';
 export { default as MGamepadList } from './MGamepadList.vue';
 export { default as MPatternList } from './MPatternList.vue';
+export { default as MAnnouncement } from './MAnnouncement.vue';
 export { default as MDiagnosticPanel } from './MDiagnosticPanel.vue';
