+++ src/models/index.ts
@@ -1,5 +1,5 @@
 export * from './common';
-export * from './Announce';
+export * from './Router';
 export * from './Pattern';
 export * from './Vibrator';
-export * from './Router';
+export * from './Announcement';
